todo review: int -> byte (beginning)

+++ HarmonyDB.Index/HarmonyDB.Index.Analysis/Models/V1/HarmonyDataV1.cs
@@ -4,7 +4,7 @@ namespace HarmonyDB.Index.Analysis.Models.V1;
 
 public record HarmonyDataV1
 {
-    public required int Root { get; init; }
+    public required byte Root { get; init; }
 
     public NoteAlteration? Alteration { get; init; }
